pyproject.toml
[tool.poetry]
name = "markup-based-segmentation-of-mna-agreements"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "markup_based_segmentation_of_mna_agreements"}]

[tool.poetry.dependencies]
python = "^3.10"
torchvision = "^0.15.2"
evaluate = "^0.4.0"
transformers = "^4.29.2"
pandas = "^2.0.2"


[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
ruff = "^0.0.270"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
